requirements.txt
Flask
python-telegram-bot[webhooks]
python-dotenv
requests
pandas
